Fix C library in global namespace deduction.
The decision to declare _STLP_VENDOR_GLOBAL_CSTD is determined
by looking for _STLP_HAS_NO_NEW_C_HEADERS.  In the situation
where new C headers are present, but are not used
(_STLP_USE_NEW_C_HEADERS is not defined), this will mistakenly
believe that the C library objects are not in the global
namespace.  The logic was changed to look at
!defined(_STLP_USE_NEW_C_HEADERS) rather than
defined(_STLP_HAS_NO_NEW_C_HEADERS).